requirements.txt
nltk==3.8.1
numpy==2.0.0
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.0
streamlit==1.36.0
